# src/docs/swagger.yaml
basePath: /
definitions:
...
  main.APICarRef:
    properties:
      car_id:
        description: CarID TeslaMate cars.id
        type: integer
      car_name:
        description: CarName 车辆昵称，可为空
        type: string
    type: object
  main.APICarsListCar:
    properties:
      car_details:
        allOf:
        - $ref: '#/definitions/main.APICarsListCarDetails'
        description: CarDetails 车型与效率等
      car_exterior:
        allOf:
        - $ref: '#/definitions/main.APICarsListCarExterior'
        description: CarExterior 外观
      car_id:
        description: CarID 车辆 ID
        type: integer
      car_settings:
        allOf:
        - $ref: '#/definitions/main.APICarsListCarSettings'
        description: CarSettings 车辆设置
      name:
        description: Name 昵称
        type: string
      teslamate_details:
        allOf:
        - $ref: '#/definitions/main.APICarsListTMDetails'
        description: TeslaMateDetails 写入/更新时间
      teslamate_stats:
        allOf:
        - $ref: '#/definitions/main.APICarsListTMStats'
        description: TeslaMateStats 充电/驾驶/更新次数统计
    type: object
...
  main.APIChargeDetail:
    properties:
      address:
        type: string
      battery_details:
        $ref: '#/definitions/main.APIChargesBatteryLevels'
      charge_details:
        items:
          $ref: '#/definitions/main.APIChargeDetailRowFixed'
        type: array
      charge_energy_added:
        type: number
      charge_energy_used:
        type: number
      charge_id:
        type: integer
      cost:
        type: number
      duration_min:
        type: integer
      duration_str:
        type: string
      end_date:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      odometer:
        type: number
      outside_temp_avg:
        type: number
      range_ideal:
        $ref: '#/definitions/main.APIChargesRangeSpan'
      range_rated:
        $ref: '#/definitions/main.APIChargesRangeSpan'
      start_date:
        type: string
    type: object
...
  main.APIChargesRow:
    properties:
      address:
        type: string
      battery_details:
        $ref: '#/definitions/main.APIChargesBatteryLevels'
      charge_energy_added:
        type: number
      charge_energy_used:
        type: number
      charge_id:
        type: integer
      cost:
        type: number
      duration_min:
        type: integer
      duration_str:
        type: string
      end_date:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      odometer:
        type: number
      outside_temp_avg:
        type: number
      range_ideal:
        $ref: '#/definitions/main.APIChargesRangeSpan'
      range_rated:
        $ref: '#/definitions/main.APIChargesRangeSpan'
      start_date:
        type: string
    type: object
  main.APIDatabaseInfo:
    properties:
      postgres_version:
        type: string
      shared_buffers_setting:
        type: string
      table_row_counts:
        items:
          $ref: '#/definitions/main.APIDatabaseTableRows'
        type: array
      table_row_counts_error:
        type: string
      table_sizes:
        items:
          $ref: '#/definitions/main.APIDatabaseTableSize'
        type: array
      timezone:
        type: string
      user_tables_total_bytes:
        type: integer
    type: object
...
  main.APIStateRow:
    properties:
      end_date:
        type: string
      id:
        type: integer
      start_date:
        type: string
      state:
        type: string
    type: object
  main.APIUnitsLengthTemp:
    properties:
      unit_of_length:
        description: UnitsLength km 或 mi
        type: string
      unit_of_temperature:
        description: UnitsTemperature C 或 F
        type: string
    type: object
  main.APIUnitsLengthTempPressure:
    properties:
      unit_of_length:
        description: UnitsLength km 或 mi
        type: string
      unit_of_pressure:
        description: UnitsPressure bar 或 psi
        type: string
      unit_of_temperature:
        description: UnitsTemperature C 或 F
        type: string
    type: object
...
  main.MQTTStatusPayload:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      status:
        $ref: '#/definitions/main.MQTTStatusSnapshot'
      units:
        $ref: '#/definitions/main.APIUnitsLengthTempPressure'
    type: object
  main.MQTTStatusResponse:
    properties:
      data:
        $ref: '#/definitions/main.MQTTStatusPayload'
    type: object
...
  main.RespAPIError:
    properties:
      error:
        description: Error 错误说明
        type: string
    type: object
...
  main.RespCarsList:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/main.RespCarsListData'
        description: Data 车辆列表或单车
    type: object
  main.RespCarsListData:
    properties:
      cars:
        description: Cars 车辆数组
        items:
          $ref: '#/definitions/main.APICarsListCar'
        type: array
    type: object
  main.RespChargeCurrent:
    properties:
      data:
        $ref: '#/definitions/main.RespChargeCurrentData'
    type: object
  main.RespChargeCurrentData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      charge:
        $ref: '#/definitions/main.APIChargeCurrent'
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
  main.RespChargeDetail:
    properties:
      data:
        $ref: '#/definitions/main.RespChargeDetailData'
    type: object
  main.RespChargeDetailData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      charge:
        $ref: '#/definitions/main.APIChargeDetail'
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
  main.RespChargesList:
    properties:
      data:
        $ref: '#/definitions/main.RespChargesListData'
    type: object
  main.RespChargesListData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      charges:
        items:
          $ref: '#/definitions/main.APIChargesRow'
        type: array
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
  main.RespDatabase:
    properties:
      data:
        $ref: '#/definitions/main.APIDatabaseInfo'
    type: object
  main.RespDriveDetail:
    properties:
      data:
        $ref: '#/definitions/main.RespDriveDetailData'
    type: object
  main.RespDriveDetailData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      drive:
        $ref: '#/definitions/main.APIDriveDetail'
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
  main.RespDrivesList:
    properties:
      data:
        $ref: '#/definitions/main.RespDrivesListData'
    type: object
  main.RespDrivesListData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      drives:
        items:
          $ref: '#/definitions/main.APIDriveListRow'
        type: array
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
...
  main.RespPositionsData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      positions:
        items:
          $ref: '#/definitions/main.APIPositionRow'
        type: array
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
  main.RespPositionsList:
    properties:
      data:
        $ref: '#/definitions/main.RespPositionsData'
    type: object
...
  main.RespStatesData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      states:
        items:
          $ref: '#/definitions/main.APIStateRow'
        type: array
    type: object
  main.RespStatesList:
    properties:
      data:
        $ref: '#/definitions/main.RespStatesData'
    type: object
  main.RespUpdatesData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      updates:
        items:
          $ref: '#/definitions/main.APIUpdateRow'
        type: array
    type: object
  main.RespUpdatesList:
    properties:
      data:
        $ref: '#/definitions/main.RespUpdatesData'
    type: object
host: localhost:8080
info:
  contact: {}
  description: 与 TeslaMate 采集数据对齐的 JSON API；看板对照见仓库 `执行计划.md`。
  license:
    name: MIT
  title: TeslaMate API
  version: "1.0"
paths:
...
  /api/v1/cars/{CarID}:
    get:
      parameters:
      - description: 车辆 ID；省略则列出全部
        example: 1
        in: path
        name: CarID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespCarsList'
      summary: 车辆列表或详情
      tags:
      - cars
...
  /api/v1/cars/{CarID}/charges:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 开始时间 RFC3339
        in: query
        name: startDate
        type: string
      - description: 结束时间 RFC3339
        in: query
        name: endDate
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 100
        description: 每页条数
        in: query
        name: show
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespChargesList'
      summary: 充电记录列表
      tags:
      - charges
  /api/v1/cars/{CarID}/charges/{ChargeID}:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 充电过程 ID
        in: path
        name: ChargeID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespChargeDetail'
      summary: 充电详情
      tags:
      - charges
  /api/v1/cars/{CarID}/charges/current:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespChargeCurrent'
      summary: 当前充电
      tags:
      - charges
  /api/v1/cars/{CarID}/command:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: POST /command/*、/wake_up：Tesla API 原始 JSON
          schema:
            items:
              type: integer
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.RespAPIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.RespAPIError'
      summary: 车辆指令 / 唤醒
      tags:
      - commands
  /api/v1/cars/{CarID}/command/{Command}:
    post:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 指令名（POST /command/:Command）
        in: path
        name: Command
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: POST /command/*、/wake_up：Tesla API 原始 JSON
          schema:
            items:
              type: integer
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.RespAPIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.RespAPIError'
      summary: 车辆指令 / 唤醒
      tags:
      - commands
  /api/v1/cars/{CarID}/drives:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 开始时间 RFC3339
        in: query
        name: startDate
        type: string
      - description: 结束时间 RFC3339
        in: query
        name: endDate
        type: string
      - description: 最小距离
        in: query
        name: minDistance
        type: number
      - description: 最大距离
        in: query
        name: maxDistance
        type: number
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 100
        description: 每页条数
        in: query
        name: show
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespDrivesList'
      summary: 行程列表
      tags:
      - drives
  /api/v1/cars/{CarID}/drives/{DriveID}:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 行程 ID
        in: path
        name: DriveID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespDriveDetail'
      summary: 行程详情
      tags:
      - drives
  /api/v1/cars/{CarID}/logging:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: PUT：TeslaMate 日志接口原始 JSON
          schema:
            items:
              type: integer
            type: array
      summary: TeslaMate 日志控制
      tags:
      - commands
  /api/v1/cars/{CarID}/logging/{Command}:
    put:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 子命令
        in: path
        name: Command
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PUT：TeslaMate 日志接口原始 JSON
          schema:
            items:
              type: integer
            type: array
      summary: TeslaMate 日志控制
      tags:
      - commands
...
  /api/v1/cars/{CarID}/positions:
    get:
      description: 查询 `positions` 表（时间序列与经纬度等）。支持 startDate、endDate 过滤 `date` 列，分页参数
        page、show（单页最大 5000）。
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 开始时间 RFC3339
        in: query
        name: startDate
        type: string
      - description: 结束时间 RFC3339
        in: query
        name: endDate
        type: string
      - default: 1
        description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - default: 500
        description: 每页条数
        in: query
        name: show
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespPositionsList'
      summary: 车辆位置与采样点
      tags:
      - dashboards
  /api/v1/cars/{CarID}/states:
    get:
      description: 查询 TeslaMate `states` 表（offline / asleep / online）。支持 startDate、endDate（RFC3339），默认按
        start_date 降序。
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 开始时间 RFC3339
        in: query
        name: startDate
        type: string
      - description: 结束时间 RFC3339
        in: query
        name: endDate
        type: string
      - default: 1
        description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - default: 100
        description: 每页条数
        in: query
        name: show
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespStatesList'
      summary: 车辆在线状态历史
      tags:
      - dashboards
  /api/v1/cars/{CarID}/status:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MQTTStatusResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/main.RespAPIError'
      summary: 实时状态（MQTT）
      tags:
      - cars
  /api/v1/cars/{CarID}/updates:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespUpdatesList'
      summary: 软件更新记录
      tags:
      - cars
  /api/v1/cars/{CarID}/wake_up:
    post:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: POST /command/*、/wake_up：Tesla API 原始 JSON
          schema:
            items:
              type: integer
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.RespAPIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.RespAPIError'
      summary: 车辆指令 / 唤醒
      tags:
      - commands
  /api/v1/database:
    get:
      description: 对齐 `database-info.json` 中与 PG 实例相关的面板。`includeTableStats=true`
        时查询 `pg_statio_user_tables`；`includeTableRows=true` 时对每表 count（可能很慢）。
      parameters:
      - default: false
        description: 是否包含表大小统计
        in: query
        name: includeTableStats
        type: boolean
      - default: false
        description: 是否估算各表行数（慢）
        in: query
        name: includeTableRows
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespDatabase'
      summary: 数据库实例信息
      tags:
      - dashboards
...
schemes:
- http
swagger: "2.0"